Move ECS code to a separate module
