+++ src/rules/infrastructure/ecs/attribute.rs
@@ -19,12 +19,12 @@ impl AttributeValue for u32 {}
 /// Each attribute has a name and value type.  For example we can
 /// create an attribute called speed that stores a u32
 /// ```
-/// # use tank_game::rules::infrastructure::attribute::Attribute;
+/// # use tank_game::rules::infrastructure::ecs::Attribute;
 /// let speed = Attribute::<u32>::new("speed");
 /// ```
 /// or define an attribute that holds a struct
 /// ```
-/// # use tank_game::rules::infrastructure::attribute::{Attribute,AttributeValue};
+/// # use tank_game::rules::infrastructure::ecs::{Attribute,AttributeValue};
 /// #[derive(Debug)]
 /// enum PetType {
 ///     Cat,
@@ -80,8 +80,8 @@ impl<ValueType: AttributeValue> Attribute<ValueType> {
 #[macro_export]
 macro_rules! attribute {
     ($name:ident: $type:ty) => {
-        pub static $name: $crate::rules::infrastructure::attribute::Attribute<$type> =
-            $crate::rules::infrastructure::attribute::Attribute::new(stringify!($name));
+        pub static $name: $crate::rules::infrastructure::ecs::Attribute<$type> =
+            $crate::rules::infrastructure::ecs::Attribute::new(stringify!($name));
     };
 }
