+++ src/rules/infrastructure/ecs/mod.rs
@@ -1,0 +1,9 @@
+mod attribute;
+mod container;
+mod pool;
+mod transaction;
+
+pub use attribute::{Attribute, AttributeValue};
+pub use container::AttributeContainer;
+pub use pool::{Pool, Index, GatheredResult, Handle};
+pub use transaction::*;
