+++ src/rules/infrastructure/ecs/transaction.rs
@@ -101,8 +101,7 @@ impl Transaction {
 /// Add the modifications required to create and initialize an attribute container to the given transaction
 ///
 /// ```
-/// # use tank_game::rules::infrastructure::attribute::Attribute;
-/// # use tank_game::rules::infrastructure::transaction::Transaction;
+/// # use tank_game::rules::infrastructure::ecs::{Attribute, Transaction};
 /// # use tank_game::create_container;
 /// # static dummy_attribute: Attribute<u32> = Attribute::<u32>::new("dummy_attribute");
 /// #
@@ -117,9 +116,9 @@ macro_rules! create_container {
         {
             use $crate::modify_container;
 
-            let transaction: &mut $crate::rules::infrastructure::transaction::Transaction = $transaction;
+            let transaction: &mut $crate::rules::infrastructure::ecs::Transaction = $transaction;
 
-            let (handle, new_container_modification) = $crate::rules::infrastructure::transaction::AddContainerModification::new();
+            let (handle, new_container_modification) = $crate::rules::infrastructure::ecs::AddContainerModification::new();
             transaction.add(new_container_modification);
 
             modify_container!(transaction, handle, {
@@ -136,8 +135,7 @@ macro_rules! create_container {
 /// A helper for creating modifications to an attribute container
 ///
 /// ```
-/// # use tank_game::rules::infrastructure::attribute::Attribute;
-/// # use tank_game::rules::infrastructure::transaction::Transaction;
+/// # use tank_game::rules::infrastructure::ecs::{Transaction, Attribute};
 /// # use tank_game::{create_container,modify_container};
 /// # static dummy_attribute: Attribute<u32> = Attribute::<u32>::new("dummy_attribute");
 /// #
@@ -151,10 +149,10 @@ macro_rules! create_container {
 macro_rules! modify_container {
     ($transaction:expr, $handle:expr, { $($attribute:ident = $value:expr),+ }) => {
         {
-            let transaction: &mut $crate::rules::infrastructure::transaction::Transaction = $transaction;
+            let transaction: &mut $crate::rules::infrastructure::ecs::Transaction = $transaction;
 
             $(
-                transaction.add($crate::rules::infrastructure::transaction::AttributeModification::new($handle, &$attribute, $value));
+                transaction.add($crate::rules::infrastructure::ecs::AttributeModification::new($handle, &$attribute, $value));
             )+
         }
     };
@@ -164,7 +162,7 @@ macro_rules! modify_container {
 mod test {
     use std::error::Error;
 
-    use crate::rules::infrastructure::{attribute::DUMMY_ATTRIBUTE, error::RuleError, pool::Index};
+    use crate::rules::infrastructure::{ecs::{attribute::DUMMY_ATTRIBUTE, pool::Index}, RuleError};
 
     use super::*;
